DRY GFQL expression source text helpers
